Split the file CGAL/Nef_3/ID_support_handler in two files, one with the generic function, one with the specialization used by Nef_3.
